conn: Return non-zero delay when in CONNECTING state
If connection is in state CONNECTING and there is data ready to be sent
using value 0 for the delay results in busy loop because Sender calls
select with the value returned from here and then just keeps on calling
it again and again until connection completes or fails. In some cases
this can take a long time and if the connection failed it will just be
retried and the busy loop continues.

This return value was changed from 999999999 to 0 in commit 16c13f91 and
it doesn't look like it needs to be exactly zero for other logic to work
correctly.

Fixes https://github.com/dpkp/kafka-python/issues/1907