Merge _find_coordinator_id methods (#2127)
Previously there were two methods:
* `_find_coordinator_id()`
* `_find_many_coordinator_ids()`

But they do basically the same thing internally. And we need the plural
two places, but the singular only one place.

So merge them, and change the function signature to take a list of
`group_ids` and return a dict of `group_id: coordinator_id`s.

As a result of this, the `describe_groups()` command should scale better
because the `_find_coordinator_ids()` command issues all the requests
async, instead of sequentially blocking as the `described_groups()` used
to do.