Bump dev requirements (#2129)
Also re-order lexicographically.

Note that I did not exhaustively test this... there could be edge cases
depending on the python version. But I think we should be okay because
`tox.ini` is currently testing using with unpinned versions, so I think
we're already running these versions in our test suite.